Add UISwitch Category to enable child switch animations
Although the app was working as designed, the child switches were not being animated when using setValue:ForKey on the array from the IBOutletCollection. Add the category to use the setOn:animated method to enable the animation. The category method is called via makeObjectsPerformSelector:withObject method on the array. This is within the rules of the challenge since there is no loop used and the method name does not begin with 'enumerate'.